% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrow.R
\name{add_arrow}
\alias{add_arrow}
\title{Add arrow to consort}
\usage{
add_arrow(consort, begin_id, end_id, direction)
}
\arguments{
\item{consort}{An object with class Consort.}

\item{begin_id}{A character with the id of the box to start the arrow.}

\item{end_id}{A character with the id of the box to end the arrow.}

\item{direction}{A character with the direction of the arrow. Use 'top' or 'side'.}
}
\value{
A consort.
}
\description{
Add arrow to consort
}
\examples{
x <- create_consort(100, 150) |>
 add_box('box1', x = c(10, 5), y = c(50, 3), align = 'c', text = 'Box 1') |>
 add_box('box2', x = c(10, 5), y = c(30, 3), align = 'c', text = 'Box 2') |>
 add_arrow('box1', 'box2', 'top')
}
