% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_box.R
\name{add_box}
\alias{add_box}
\title{Add box to consort}
\usage{
add_box(consort, id, x, y, align, text, fontsize = 8)
}
\arguments{
\item{consort}{An object with class Consort.}

\item{id}{A character with de the id of the box. Must be unique.}

\item{x}{An array with 2 elements: first element is the position in x-axis and second is the half-length of the box in x-axis.}

\item{y}{An array with 2 elements: first element is the position in y-axis and second is the half-length of the box in y-axis.}

\item{align}{A character with the alignment of the text. Use 'l' to left, 'c' to center and 'r' to right.}

\item{text}{A character with the text inside the box.}

\item{fontsize}{A numeric with the font size of text inside the box.}
}
\value{
A consort.
}
\description{
Add box to consort
}
\examples{
x <- create_consort(100, 150) |>
 add_box(id = 'box1', x = c(10, 5), y = c(50, 3), align = 'c', text = 'Box 1')
}
