% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_consort.R
\name{create_consort}
\alias{create_consort}
\title{Create consort}
\usage{
create_consort(xsize, ysize)
}
\arguments{
\item{xsize}{A number with x-axis dimension.}

\item{ysize}{A number with y-axis dimension.}
}
\value{
A consort.
}
\description{
Create consort
}
\examples{
x <- create_consort(100, 150)
}
