% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consort.R
\name{get_consort}
\alias{get_consort}
\title{Get consort plot}
\usage{
get_consort(consort)
}
\arguments{
\item{consort}{An object with class Consort.}
}
\value{
A ggplot object
}
\description{
Get consort plot
}
\examples{
x <- create_consort(100, 150) |>
 add_box('box1', x = c(10, 5), y = c(50, 3), align = 'c', text = 'Box 1') |>
 display_consort() |>
 get_consort()
}
